requirements.txt
git+https://github.com/pyLerner/folium-pat@main#egg=folium
PyMySQL==1.0.2
SQLAlchemy==1.4.45
branca==0.6.0
certifi==2022.9.24
charset-normalizer==2.1.1
dateutils==0.6.12
distlib==0.3.
et-xmlfile==1.1.0
filelock==3.8.0
gtfs-realtime-bindings==0.0.7
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.23.4
openpyxl==3.0.10
pandas==1.5.1
Pillow==9.4.0
platformdirs==2.5.3
protobuf==3.20.3
python-dateutil==2.8.2
pytz==2022.6
rarfile==4.0
requests==2.28.1
six==1.16.0
sklearn==0.0.post1
urllib3==1.26.12
virtualenv==20.16.6
virtualenvwrapper-win==1.2.7
xlrd==2.0.1
fastapi~=0.95.2
uvicorn~=0.22.0
python-multipart
aiofiles~=23.1.0

python-dotenv~=1.0.0